create helper function for event generator
