+++ examples/datawarehouse/index.ts
@@ -1,8 +1,7 @@
 import * as pulumi from "@pulumi/pulumi";
 import * as aws from "@pulumi/aws";
-import { EventRuleEvent } from "@pulumi/aws/cloudwatch";
-import { CallbackFunction } from "@pulumi/aws/lambda";
 import { ServerlessDataWarehouse, StreamingInputTableArgs } from "../../lib/datawarehouse";
+import { createEventGenerator } from "./eventGenerator";
 
 // app specific config
 const config = new pulumi.Config();
@@ -34,8 +33,8 @@ const columns = [
     }
 ];
 
-const logsTableName = "logs";
-const logsTableArgs: StreamingInputTableArgs = {
+const impressionsTableName = "impressions";
+const impressionsTableArgs: StreamingInputTableArgs = {
     columns,
     inputStreamShardCount: shards[stage],
     region,
@@ -43,74 +42,9 @@ const logsTableArgs: StreamingInputTableArgs = {
 }
 
 const dataWarehouse = new ServerlessDataWarehouse("analytics_dw")
-    .withStreamingInputTable(logsTableName, logsTableArgs);
+    .withStreamingInputTable(impressionsTableName, impressionsTableArgs);
 
-const logsInputStream = dataWarehouse.getInputStream(logsTableName);
-export const streamName = logsInputStream.name;
+const impressionsInputStream = dataWarehouse.getInputStream(impressionsTableName);
+export const streamName = impressionsInputStream.name;
 
-// generate test events
-let lambdaAssumeRolePolicy = {
-    "Version": "2012-10-17",
-    "Statement": [
-        {
-            "Action": "sts:AssumeRole",
-            "Principal": {
-                "Service": "lambda.amazonaws.com",
-            },
-            "Effect": "Allow",
-            "Sid": "",
-        },
-    ],
-};
-
-const eventGenRole = new aws.iam.Role("eventGenLambdaRole", {
-    assumeRolePolicy: JSON.stringify(lambdaAssumeRolePolicy),
-});
-
-const eventGenLambdaAccess = new aws.iam.RolePolicyAttachment("event-gen-lambda-access", {
-    role: eventGenRole,
-    policyArn: aws.iam.ManagedPolicies.AWSLambdaFullAccess,
-});
-
-const eventGenKinesisAccess = new aws.iam.RolePolicyAttachment("event-gen-kinesis-access", {
-    role: eventGenRole,
-    policyArn: aws.iam.ManagedPolicies.AmazonKinesisFullAccess,
-});
-
-
-const eventCron = new aws.cloudwatch.EventRule("event-gen-cron", {
-    scheduleExpression: "rate(1 minute)",
-});
-
-eventCron.onEvent("event-generator", new CallbackFunction('event-gen-callback', {
-    role: eventGenRole,
-    callback: (event: EventRuleEvent) => {
-        const AWS = require("aws-sdk");
-        const uuid = require("uuid/v4")
-        const kinesis = new AWS.Kinesis();
-        const records: any = [];
-
-        const sessionId = uuid();
-        const eventId = uuid();
-        const record = {
-            Data: JSON.stringify({
-                id: eventId,
-                session_id: sessionId,
-                message: "this is a message",
-                event_type: "impression",
-            }),
-            PartitionKey: sessionId
-        };
-        records.push(record);
-
-        kinesis.putRecords({
-            Records: records,
-            StreamName: streamName.get()
-        }, (err: any) => {
-            if (err) {
-                console.error(err)
-            }
-        });
-    }
-
-}));
+createEventGenerator("impression", impressionsInputStream.name.get());
